+++ game/src/main/java/com/blaj/openmetin/game/application/features/serverstatus/ServerStatusCommand.java
@@ -1,5 +1,6 @@
 package com.blaj.openmetin.game.application.features.serverstatus;
 
 import com.blaj.openmetin.shared.infrastructure.cqrs.Command;
+import org.joou.UByte;
 
-public record ServerStatusCommand(int channelIndex) implements Command<Void> {}
+public record ServerStatusCommand(UByte channelIndex) implements Command<Void> {}
